Color module renamed.
